✅ Status: changed to djs-v13, added linter, es6+ implemented
