+++ .eslintrc.json
@@ -1,0 +1,14 @@
+{
+	"extends": "eslint:recommended",
+	"env": {
+		"node": true,
+		"es6": true
+	},
+	"parserOptions": {
+		"ecmaVersion": 2021,
+        "sourceType": "module"
+	},
+	"rules": {
+
+	}
+}
