--- Procfile
@@ -1,1 +1,0 @@
-worker node src/bot.js
