+++ junk old code/backup_bot.js
