+++ junk old code/commands/help.js
