+++ junk old code/commands/ping.js
