+++ junk old code/connection.js
