+++ junk old code/nice code/voice.js
