+++ junk old code/rank.js
