+++ junk old code/ranks_.js
