+++ sample.env
@@ -1,2 +1,1 @@
-MONGOIRI=YOUR_MONGO_URI
 DJSTOKEN=DISCORD_API_TOKEN
